refactor(declaration-remuneration): #4098 — extrait la projection ligne → étapes hors des helpers
`declarationHelpers.ts` dépassait les 400 lignes après l'ajout des écarts
persistés. `mapToStepData`, qui projette une ligne de déclaration vers les
formes de formulaire des étapes 2 à 4, part dans son propre module.

Au passage, `GipMdsRow` n'est plus redéclaré : le type est importé depuis
`gipMdsMapping`, où il est déjà défini et exporté.
