+++ packages/app/src/app/declaration-remuneration/(with-banner)/etape/[step]/page.tsx
@@ -14,10 +14,8 @@ import {
 	isDeclarationSubmitted,
 	isIndicatorGRequired,
 } from "~/modules/domain";
-import {
-	mapToEmployeeCategoryRows,
-	mapToStepData,
-} from "~/server/api/routers/declarationHelpers";
+import { mapToEmployeeCategoryRows } from "~/server/api/routers/declarationHelpers";
+import { mapToStepData } from "~/server/api/routers/declarationStepMapping";
 import { getCampaignDeadlines } from "~/server/db/getCampaignDeadlines";
 import { api, HydrateClient } from "~/trpc/server";
