+++ packages/app/src/server/api/routers/adminDeclarations.ts
@@ -29,10 +29,8 @@ import {
 	parseGipWorkforce,
 	toDisplayWorkforce,
 } from "~/modules/domain";
-import {
-	mapToEmployeeCategoryRows,
-	mapToStepData,
-} from "~/server/api/routers/declarationHelpers";
+import { mapToEmployeeCategoryRows } from "~/server/api/routers/declarationHelpers";
+import { mapToStepData } from "~/server/api/routers/declarationStepMapping";
 import { adminProcedure, createTRPCRouter } from "~/server/api/trpc";
 import {
 	companies,
